refactor(composer-devtools-protocol): drop extension-side code
Per review feedback: the protocol package now contains only the spec
typedefs and the page-side client API (builders, ComposerDevtools,
installDevtools). The content-script bridge, background router, panel
connector, and chrome.runtime.Port endpoint adapter all move into
composer-crx where they belong.
