+++ packages/apps/composer-crx/src/content.ts
@@ -5,11 +5,11 @@
 import { onMessage, sendMessage } from 'webext-bridge/content-script';
 import browser from 'webextension-polyfill';
 
-import { installDevtoolsBridge } from '@dxos/composer-devtools-protocol/bridge';
 import { log } from '@dxos/log';
 
 import { CLIP_ACK_EVENT, CLIP_EVENT, type Clip, type ClipAck } from './clip/types';
 import { DEVELOPER_MODE_PROP, getProp } from './config';
+import { installDevtoolsBridge } from './devtools/bridge';
 import { pickAndHarvest } from './picker';
 import { showDebugPreview } from './picker/debug-preview';
